content/events/test/test_bug822898.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=822898
-->
<head>
  <title>Test for Bug 822898</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=822898">Mozilla Bug 822898</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.8">

/** Test for Bug 822898 - Pointer* Events **/

let tests = [], testTarget, parent;

function nextTest() {
  if (tests.length)
    SimpleTest.executeSoon(tests.shift());
}

function random() {
  return Math.floor(Math.random() * 100);
}

function createTestEventValue(name) {

  let detail = random();
  let screenX = random();
  let screenY = random();
  let clientX = random();
  let clientY = random();
  let ctrlKey = random() % 2 ? true : false;
  let altKey = random() % 2 ? true : false;
  let shiftKey = random() % 2 ? true : false;
  let metaKey = random() % 2 ? true : false;
  let button = random();
  let pointerId = random();

  return function() {
    let event = new PointerEvent("pointerdown", {
      bubbles: true, cancelable: true, view: window,
      detail: detail, screenX: screenX, screenY: screenY, clientX: clientX, clientY: clientY,
      ctrlKey: ctrlKey, altKey: altKey, shiftKey: shiftKey, metaKey: metaKey,
      button: button, relatedTarget: null, pointerId: pointerId
    });


    function check(ev) {
      is(ev.detail, detail, "Correct detail");
      is(ev.screenX, screenX, "Correct screenX");
      is(ev.screenY, screenY, "Correct screenY");
      is(ev.clientX, clientX, "Correct clientX");
      is(ev.clientY, clientY, "Correct clientY");
      is(ev.ctrlKey, ctrlKey, "Correct ctrlKey");
      is(ev.altKey, altKey, "Correct altKey");
      is(ev.shiftKey, shiftKey, "Correct shiftKey");
      is(ev.metaKey, metaKey, "Correct metaKey");
      is(ev.button, button, "Correct buttonArg");
      is(ev.pointerId, pointerId, "Correct pointerId");
    }

    for each (let target in [document, window, testTarget, parent])
      target.addEventListener(name, check, false);

    testTarget.dispatchEvent(event);

    for each (let target in [document, window, testTarget, parent])
      target.removeEventListener(name, check, false);


    nextTest();
  }
}

function getDefaultArgEvent(eventname) {
  return new PointerEvent(eventname, {
    bubbles: true, cancelable: true, view: window,
    detail: 0, screenX: 0, screenY: 0, clientX: 0, clientY: 0,
    ctrlKey: false, altKey: false, shiftKey: false, metaKey: false,
    button: 0, relatedTarget: null, pointerId: 0
  });
}

function testDefaultArg() {
  let event = getDefaultArgEvent("pointerdown");

  testTarget.addEventListener("pointerdown", function(ev) {
    testTarget.removeEventListener("pointerdown", arguments.callee, false);
    is(ev.pointerId, 0, "Correct default pointerId");
  }, false);
  testTarget.dispatchEvent(event);

  nextTest();
}

function testStopPropagation() {
  let event = getDefaultArgEvent("pointerdown");

  let unreachableListener = function () {
    ok(false, "Event should have been stopped");
  }

  // Capturing phase
  let captured = false;
  parent.addEventListener("pointerdown", function() {
    parent.removeEventListener("pointerdown", arguments.callee, true);
    captured = true;
  }, true); // Capturing phase

  // Bubbling phase
  parent.addEventListener("pointerdown", unreachableListener, false);

  testTarget.addEventListener("pointerdown", function(ev) {
    testTarget.removeEventListener("pointerdown", arguments.callee, false);
    is(captured, true, "Event should have been captured");
    ev.stopPropagation();
  }, false);

  testTarget.dispatchEvent(event);

  parent.removeEventListener("pointerdown", unreachableListener, false);

  nextTest();
}

function testPreventDefault() {
  let event = getDefaultArgEvent("pointerdown");

  parent.addEventListener("pointerdown", function(ev) {
    parent.removeEventListener("pointerdown", arguments.callee, false);
    is(ev.defaultPrevented, true, "preventDefault can be called");
    nextTest();
  }, false);

  testTarget.addEventListener("pointerdown", function(ev) {
    testTarget.removeEventListener("pointerdown", arguments.callee, false);
    ev.preventDefault();
  }, false);

  testTarget.dispatchEvent(event);
}

function testBlockPreventDefault() {
  let event = new PointerEvent("pointerdown", {
    bubbles: true, cancelable: false, view: window,
    detail: 0, screenX: 0, screenY: 0, clientX: 0, clientY: 0,
    ctrlKey: false, altKey: false, shiftKey: false, metaKey: false,
    button: 0, relatedTarget: null, pointerId: 0, pointerType: "pen"
  });

  parent.addEventListener("pointerdown", function(ev) {
    parent.removeEventListener("pointerdown", arguments.callee, false);
    is(ev.defaultPrevented, false, "aCancelableArg works");
    nextTest();
  }, false);

  testTarget.addEventListener("pointerdown", function(ev) {
    testTarget.removeEventListener("pointerdown", arguments.callee, false);
    ev.preventDefault();
  }, false);

  testTarget.dispatchEvent(event);
}

function testBlockBubbling() {
  let unreachableListener = function () {
    ok(false, "aCanBubble doesn't work");
  }

  let event = new PointerEvent("pointerdown", {
    bubbles: false, cancelable: true, view: window,
    detail: 0, screenX: 0, screenY: 0, clientX: 0, clientY: 0,
    ctrlKey: false, altKey: false, shiftKey: false, metaKey: false,
    button: 0, relatedTarget: null, pointerId: 0
  });

  parent.addEventListener("pointerdown", unreachableListener, false);
  testTarget.dispatchEvent(event);
  parent.removeEventListener("pointerdown", unreachableListener, false);

  nextTest();
}

var gOnPointerPropHandled = new Array;

function testOnPointerProperty()
{

  testTarget.onpointerdown = function (e) { gOnPointerPropHandled["pointerdown"] = true; }
  testTarget.onpointerup = function (e) { gOnPointerPropHandled["pointerup"] = true; }
  testTarget.onpointermove = function (e) { gOnPointerPropHandled["pointermove"] = true; }
  testTarget.onpointerout = function (e) { gOnPointerPropHandled["pointerout"] = true; }
  testTarget.onpointerover = function (e) { gOnPointerPropHandled["pointerover"] = true; }
  testTarget.onpointerenter = function (e) { gOnPointerPropHandled["pointerenter"] = true; }
  testTarget.onpointerleave = function (e) { gOnPointerPropHandled["pointerleave"] = true; }

  testTarget.dispatchEvent(getDefaultArgEvent("pointerdown"));
  ok(gOnPointerPropHandled['pointerdown'], "pointerdown property isn't performed");

  testTarget.dispatchEvent(getDefaultArgEvent("pointerup"));
  ok(gOnPointerPropHandled['pointerup'], "pointerup property isn't performed");

  testTarget.dispatchEvent(getDefaultArgEvent("pointermove"));
  ok(gOnPointerPropHandled['pointermove'], "pointermove property isn't performed");

  testTarget.dispatchEvent(getDefaultArgEvent("pointerout"));
  ok(gOnPointerPropHandled['pointerout'], "pointerout property isn't performed");

  testTarget.dispatchEvent(getDefaultArgEvent("pointerover"));
  ok(gOnPointerPropHandled['pointerover'], "pointerover property isn't performed");

  testTarget.dispatchEvent(getDefaultArgEvent("pointerenter"));
  ok(gOnPointerPropHandled['pointerenter'], "pointerenter property isn't performed");

  testTarget.dispatchEvent(getDefaultArgEvent("pointerleave"));
  ok(gOnPointerPropHandled['pointerleave'], "pointerleave property isn't performed");

  nextTest();
}


function doTest() {
  SpecialPowers.setBoolPref("dom.w3c_pointer_events.enabled", true);      // Enable Pointer Events
  testTarget = document.getElementById("testTarget");
  parent = testTarget.parentNode;

  tests.push(createTestEventValue("pointerdown"));
  tests.push(createTestEventValue("pointermove"));
  tests.push(createTestEventValue("pointerup"));

  tests.push(testDefaultArg);
  tests.push(testStopPropagation);

  tests.push(testPreventDefault);
  tests.push(testBlockPreventDefault);

  tests.push(testBlockBubbling);
  tests.push(testOnPointerProperty());

  tests.push(function() {
    SpecialPowers.clearUserPref("dom.w3c_pointer_events.enabled");      // Disable Pointer Events
    SimpleTest.finish();
  });

  nextTest();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(doTest);

</script>
</pre>
<div id="parent">
  <span id="testTarget" style="border: 1px solid black;">testTarget</span>
</div>
</body>
</html>
